requirements3.txt
docutils==0.14
docopt==0.6.2
pytest==4.4.1
networkx==2.1
ply==3.11
tqdm==4.31.1
nltk==3.4.1
bioc==1.3.1
pystanforddependencies==0.3.1
bllipparser==2016.9.11
pymetamap==0.1
JPype1==0.6.3
numpy==1.16.4
pytest-cov